--- a/scripts/lib/CIME/XML/compilerblock.py
+++ b/scripts/lib/CIME/XML/compilerblock.py
@@ -114,7 +114,7 @@ class CompilerBlock(object):
         modified and thus serve as additional outputs.
         """
         writer = self._writer
-        output = elem.text
+        output = elem.text()
         if output is None:
             output = ""
         logger.debug("Initial output={}".format(output))
@@ -159,7 +159,7 @@ class CompilerBlock(object):
         for child in elem:
             if child.tag == "env":
                 # <env> tags just need to be expanded by the writer.
-                output += writer.environment_variable_string(child.text)
+                output += writer.environment_variable_string(child.text())
             elif child.tag == "shell":
                 # <shell> tags can contain other tags, so handle those.
                 command = self._handle_references(child, set_up, tear_down,
@@ -175,7 +175,7 @@ class CompilerBlock(object):
             elif child.tag == "var":
                 # <var> commands also need expansion by the writer, and can
                 # add dependencies.
-                var_name = child.text
+                var_name = child.text()
                 output += writer.variable_string(var_name)
                 depends.add(var_name)
             else: