--- a/scripts/lib/CIME/XML/compilers.py
+++ b/scripts/lib/CIME/XML/compilers.py
@@ -237,7 +237,7 @@ class Compilers(GenericXML):
 
 def _add_to_macros(node, macros):
     for child in node:
-        name = child.tag
+        name = child.tag()
         attrib = child.attrib()
         value = child.text()