--- a/scripts/lib/CIME/XML/env_base.py
+++ b/scripts/lib/CIME/XML/env_base.py
@@ -150,7 +150,7 @@ class EnvBase(EntryID):
                 # now we seperated into component and compclass)
                 if len(componentatt) > 0:
                     for ccnode in vnode.findall(".//value[@component]"):
-                        val = ccnode.attrib.get("component")
+                        val = ccnode.get("component")
                         ccnode.pop("component")
                         ccnode.set("compclass", val)
                 compclassatt = vnode.findall(".//value[@compclass]")