--- a/scripts/lib/CIME/XML/env_batch.py
+++ b/scripts/lib/CIME/XML/env_batch.py
@@ -593,7 +593,7 @@ class EnvBatch(EnvBase):
             logger.warning("Batch queries not supported on this platform")
         else:
             cmd = batch_query.text + " "
-            if "per_job_arg" in batch_query.attrib:
+            if batch_query.has("per_job_arg"):
                 cmd += batch_query.get("per_job_arg") + " "
 
             cmd += jobid