Hi all,

I wanted to start a discussion about CIME's performance and it's use of XML as a database. I've been doing some unrelated work for ACME where I've been watching "top" while running create_test. Far too often, I see cases bottle-necked in python instead of building/running. Some recent testing on melvin (fast machine, local disk) consistently shows that it takes about 15 seconds for the first 3 phases (create, xml, setup) on the cime_tiny tests, with case_setup being the most expensive at about 10 seconds. Surely, we can do better than this?

I profiled case.setup and attached the profile to our slack. The thing that caught my eye is that we're spending 60% of our runtime in generic_xml.get_nodes, so this is obviously the area to focus  on. I do not have much experience with python XML libraries, but it seems like we're being very inefficient in the way we're doing queries.

In this PR, I did some prototyping. The goal was to fully encapsulate our use of XML (no direct calls to XML libs) and to try to use caching to speed things up. Our encapsulation was pretty good with most queries going through our classes, but direct access of XML elements was done in several places.

I had some modest success with this toy PR, reducing time spent in get_nodes by 50%, and reducing the cost of case.submit from 10 seconds down to 7 seconds on melvin. Logging showed about 10k cache "hits" and 5k "misses".

Anyway, I don't trust this PR and this was only intended as a discussion topic. Ideally, I'd like to see full encapsulation of our use of XML, and the addition of a high-performing internal representation as an in-memory intermediary between CIME and the XML file database.