@billsacks 2017-12-12
This seems like a very worthwhile effort. I know very little about this part of cime, so I'll defer to others in the review. One question, though: Before we start doing our own optimizations: Does anyone know if we're using an optimal xml parser, or if there is a more optimized version out there? e.g., some googling suggests using `cElementTree` rather than `ElementTree`. 

I tried simply

```diff
diff --git a/scripts/lib/CIME/XML/standard_module_setup.py b/scripts/lib/CIME/XML/standard_module_setup.py
index 8ab608c5d..8f4cc5842 100644
--- a/scripts/lib/CIME/XML/standard_module_setup.py
+++ b/scripts/lib/CIME/XML/standard_module_setup.py
@@ -4,7 +4,7 @@
 import logging
 import os
 import sys
-import xml.etree.ElementTree as ET
+import xml.etree.cElementTree as ET
 import re
 LIB_DIR = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 sys.path.append(LIB_DIR)
```

but apparently we'd need to do a bit more than simply that.