@jgfouca 2017-12-12
@billsacks , that's a good idea and one of the first things I tried was exactly what you did. It seems in more recent versions of python, cElementTree and ElementTree have identical performance.