@jgfouca 2017-12-12
@jedwards4b yes, we can certainly do better in terms of design. The real solution would probably use GenericXML as the interface to the database as you suggested. Don't worry too much about the details of this PR, it will never be merged.