@jgfouca 2018-01-08
Replaced with #2148 and #2205